adding accountId queries to support multiple accounts

+++ Alice/SessionRecord.cpp
@@ -3,12 +3,13 @@
 using namespace sqlite;
 using namespace std;
 
-CriptextDB::SessionRecord CriptextDB::getSessionRecord(database db, string recipientId, long int deviceId) {
+CriptextDB::SessionRecord CriptextDB::getSessionRecord(database db, int accountId, string recipientId, long int deviceId) {
   string myRecord = "";
   int myLen = 0;
-  db << "Select * from sessionrecord where recipientId == ? and deviceId == ?;"
+  db << "Select * from sessionrecord where recipientId == ? and deviceId == ? and accountId == ?;"
      << recipientId
      << deviceId
+     << accountId
      >> [&] (string recipientId, int deviceId, string record, int recordLength) {
         myLen = recordLength;
 		myRecord = record;
@@ -26,11 +27,12 @@ CriptextDB::SessionRecord CriptextDB::getSessionRecord(database db, string recip
   return sessionRecord;
 }
 
-vector<CriptextDB::SessionRecord> CriptextDB::getSessionRecords(database db, string recipientId) {
+vector<CriptextDB::SessionRecord> CriptextDB::getSessionRecords(database db, int accountId, string recipientId) {
   vector<CriptextDB::SessionRecord> sessionRecords;
   try {
-	  db << "Select * from sessionrecord where recipientId == ?;"
+	  db << "Select * from sessionrecord where recipientId == ? and accountId == ?;"
      << recipientId
+     << accountId
      >> [&] (string recipientId, int deviceId, string record, int recordLength) {
         SessionRecord mySessionRecord = { 
           recipientId, 
@@ -48,29 +50,31 @@ vector<CriptextDB::SessionRecord> CriptextDB::getSessionRecords(database db, str
   return sessionRecords;
 }
 
-bool CriptextDB::createSessionRecord(database db, string recipientId, long int deviceId, char* record, size_t len) {
+bool CriptextDB::createSessionRecord(database db, int accountId, string recipientId, long int deviceId, char* record, size_t len) {
   try {
   	bool hasRow = false;
     db << "begin;";
-    db << "Select * from sessionrecord where recipientId == ? and deviceId == ?;"
+    db << "Select * from sessionrecord where recipientId == ? and deviceId == ? and accountId == ?;"
      << recipientId
      << deviceId
+     << accountId
      >> [&] (string recipientId, int deviceId, string record, int recordLength) {
         hasRow = true;
     };
-    std::cout << 20 << std::endl;
     if (hasRow) {
-      db << "update sessionrecord set record = ?, recordLength = ? where recipientId == ? and deviceId == ?;"
+      db << "update sessionrecord set record = ?, recordLength = ? where recipientId == ? and deviceId == ? and accountId == ?;"
         << record
         << static_cast<int>(len)
         << recipientId
-        << deviceId;
+        << deviceId
+        << accountId;
     } else {
-      db << "insert into sessionrecord (recipientId, deviceId, record, recordLength) values (?,?,?,?);"
+      db << "insert into sessionrecord (recipientId, deviceId, record, recordLength, accountId) values (?,?,?,?,?);"
         << recipientId
         << deviceId
         << record
-        << static_cast<int>(len);
+        << static_cast<int>(len)
+        << accountId;
     }
     db << "commit;";
   } catch (exception& e) {
@@ -80,11 +84,12 @@ bool CriptextDB::createSessionRecord(database db, string recipientId, long int d
   return true;
 }
 
-bool CriptextDB::deleteSessionRecord(database db, string recipientId, long int deviceId) {
+bool CriptextDB::deleteSessionRecord(database db, int accountId, string recipientId, long int deviceId) {
   try {
-    db << "delete from sessionrecord where recipientId == ? and deviceId == ?;"
-     << recipientId
-     << deviceId;
+    db << "delete from sessionrecord where recipientId == ? and deviceId == ? and accountId == ?;"
+      << recipientId
+      << deviceId
+      << accountId;
   } catch (exception& e) {
     std::cout << "ERROR : " << e.what() << std::endl;
     return false;
@@ -92,10 +97,11 @@ bool CriptextDB::deleteSessionRecord(database db, string recipientId, long int d
   return true;
 }
 
-bool CriptextDB::deleteSessionRecords(database db, string recipientId) {
+bool CriptextDB::deleteSessionRecords(database db, int accountId, string recipientId) {
   try {
-    db << "delete from sessionrecord where recipientId == ?;"
-     << recipientId;
+    db << "delete from sessionrecord where recipientId == ? and accountId == ?;"
+      << recipientId
+      << accountId;
   } catch (exception& e) {
     std::cout << "ERROR : " << e.what() << std::endl;
     return false;
